Respect user-supplied quoting
Currently it's impossible to have 'lazily' evaluated environment
variables like

```
  FOO='${BAR}'
```

where `${BAR}` is evaluated at runtime.

This is because we strip any quotes (single or double) and replace them
with double quotes, and then `${BAR}` is expanded.

Instead, we can check if the value is already quoted, and only supply
our own double quotes if it's not.